### 2025-12-03 #9175 @olexandr-konovalov
Hello @mdavezac @StewMH @gomezzz @theo-barfoot - I am looking for two reviewers for this paper in JOSS. Would some of you be interested and able to review it, please?
