### 2025-12-16 #9175 @olexandr-konovalov
@tesch1 @ianhbell @AndrewGibbs @pregnault @bsun1220

I am looking for two reviewers for this paper in JOSS. Would some of you be interested and able to review it, please?
