### 2026-01-18 #9175 @olexandr-konovalov
Hello @drvinceknight, @felipefr, @scarlehoff, @ziyiyin97, @timokoch - I am looking for two reviewers for this paper in JOSS. Would some of you be interested and able to review it, please?
