### 2026-01-26 #9175 @scarlehoff
fwiw I'd be able to review, but not in the next two weeks as I have a few deadlines looming.
