### 2026-01-26 #9905 @editorialbot
[REVIEW]: pyaudi: A truncated Taylor polynomial algebra toolbox for differentiable intelligence, automatic differentiation, and verified integration applications
**Submitting author:** @sbcow (<a href="http://orcid.org/0009-0000-9912-9354">Sean Cowan</a>)
**Repository:** https://github.com/darioizzo/audi
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.10.0
