### 2026-02-10 #9905 @scarlehoff
- [x] **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item. - https://github.com/darioizzo/audi/blob/paper/paper/paper_results.md

  - [ ] Trying to follow the quickstart guide https://github.com/darioizzo/audi/issues/70#issue-3920888105

- [ ] **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support **MISSING**
