### 2026-02-10 darioizzo/audi#70 @scarlehoff
Issue when trying to use the quickstart examples with conda installation
I'm following the quickstart guide in the documentation https://darioizzo.github.io/audi/quickstart.html

I've installed audi and pyaudi from conda-forge (using pixi) and, while the python code works out of the box (as expected), the C++ code needs a few extra steps.

While the "from source" installation mentions the dependencies, they are not included in the "package" documentation and are not included in the conda forge package (I see they are used at built time in the feedstock though, perhaps they should be added to `run`:
https://github.com/conda-forge/audi-feedstock/blob/442937f24564fa9e7e36213d530a546ff0b6366c/recipe/meta.yaml#L23)

So, with that in mind, I tried the two following steps to test it:

1. Add `gxx` to my conda environment (to avoid using the system g++ and ensure that it can automagically find the audi headers)
2. Add the dependencies `obake-devel libboost-devel eigen mppp` from conda-forge

With that however, the example still doesn't compile. Since the error seem to be coming from some of the dependencies (since I'm getting the latest versions) I tried to either use `std=g++14` or to downgrade some of the libraries but didn't seem to work.

I think it would be nice to add which versions of the libraries `audi` is compatible with (ideally if they can be installed with conda-forge, please add also the exact version).

Note: due to some limitations in licensing I can only install from `conda-forge` at the moment. Let me know if I tried to use the default anaconda repos instead.
