### 2026-03-16 #9905 @scarlehoff
> @scarlehoff thank you for starting the review. Just wanted to ask if you have managed to continue with further checks, or plan to do that soon.

I am waiting for feedback from the authors on this issue https://github.com/darioizzo/audi/issues/70#issue-3920888105 to be able to use the quickstart examples and continue the review.
