### 2026-03-16 #9905 @sbcow
I was unaware that this was the bottleneck, apologies. We will look into it in the next few days and come back to you on that issue
