### 2026-03-18 darioizzo/audi#70 @sbcow
Hey @scarlehoff, may I ask what error you got exactly?

Edit: And, were you able to compile audi from source and run the ctest suite successfully?

Edit2:
One thing to check is that Eigen in your environment is <5.0.0 (and >=3.0.0), which does not happen automatically if you retrieve the latest. (I will update the environment.yml accordingly)
The `g++ -std=c++11 getting_started.cpp -lmpfr -lgmp -pthread` command will not work out of the box, as it depends on a number of things on one's system. After compiling audi from source and having all dependencies in an `audi` conda environment, I was able to run the quickstart cpp code (on macos) by running:
```
clang++ audi_test.cpp \
  -std=c++20 \
  -I$CONDA_PREFIX/include \
  -L$CONDA_PREFIX/lib \
  -lmpfr -lgmp \
  -lboost_system \
  -lobake \
  -ltbb \
  -lmp++ \
  -lfmt \
  -pthread \
  -o audi_test
```

On linux, the same command worked for me using `g++`.

When trying the same command with audi installed from conda-forge, it fails I _think_ because the audi package from conda doesn't ship obake with the headers, but only the library.
