### 2026-03-19 darioizzo/audi#70 @scarlehoff
I'm using `conda-forge` indeed, the default repositories of conda can be a bit messy with licenses and so... below is the `pixi` environment.

eigen seems to be correct.

After playing a bit with it, I think most of the errors are linker errors related to abseil:

```
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: /tmp/ccFfie6x.o: in function `absl::lts_20240116::container_internal::AssertIsFull(absl::lts_20240116::container_internal::ctrl_t const*, unsigned char, unsigned char const*, char const*)':
1618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc[_ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc]+0x6f): undefined reference to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)'
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: getting_started.cpp:(.text._ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc[_ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc]+0xe7): undefined reference to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)'
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: getting_started.cpp:(.text._ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc[_ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc]+0x168): undefined reference to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)'
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: getting_started.cpp:(.text._ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc[_ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc]+0x1e9): undefined reference to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)'
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: getting_started.cpp:(.text._ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc[_ZN4absl12lts_2024011618container_internal12AssertIsFullEPKNS1_6ctrl_tEhPKhPKc]+0x266): undefined reference to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)'
audi/review/.pixi/envs/default/bin/../libexec/gcc/x86_64-conda-linux-gnu/15.2.0/ld: /tmp/ccFfie6x.o:getting_started.cpp:(.text._ZN4absl12lts_2024011618container_internal26AssertIsValidForComparisonEPKNS1_6ctrl_tEhPKh[_ZN4absl12lts_2024011618container_internal26AssertIsValidForComparisonEPKNS1_6ctrl_tEhPKh]+0xa9): more undefined references to `absl::lts_20240116::raw_log_internal::RawLog(absl::lts_20240116::LogSeverity, char const*, int, char const*, ...)' follow
```

but I really don't want to chase down the right libraries to link. Perhaps a conda-forge installation is not advisable, I'll try again once I'm in a computer in which I won't get sued if I use the anaconda repository...

```
[workspace]
channels = ["conda-forge"]
name = "review"
platforms = ["linux-64"]
version = "0.1.0"

[tasks]

[dependencies]
pyaudi = ">=1.9.2,<2"
audi = ">=1.9.2,<2"
obake-devel = ">=0.8.0,<0.9"
libboost-devel = "==1.82"
eigen = ">=3.4.0,<4"
mppp = ">=1.0.2,<2"
gxx = ">=15.2.0,<15.3"
obake = ">=0.8.0,<0.9"
```
