### 2026-03-31 #9905 @tesch1
I see the separation of the `paper_results.md` happened in pre-review, which is almost unfortunate, but the result is a raw `.md` file with latex which isn't rendered by GitHub.  Technically the data is accessible, checking the box, but I wish it would render.
