### 2026-04-02 darioizzo/audi#71 @tesch1
contribution guidelines?
Are there any contribution guidelines for this project?
