### 2026-04-02 darioizzo/audi#71 @sbcow
At this point there are none. If needed we can add templates for issues and a CONTRIBUTING.md
