### 2026-04-04 darioizzo/audi#71 @tesch1
https://github.com/openjournals/joss-reviews/issues/9905#issuecomment-4064708005

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
