### 2026-05-14 darioizzo/audi#70 @scarlehoff
I can confirm with the updated docs I can now follow the quickstart guide without (almost) no extra steps.

I still had to install `gxx libboost-devel obake-devel`, which you might want to add explicitly to the documentation here https://darioizzo.github.io/audi/install_c.html#conda for completeness given that you have the explicit conda subsection.
